Commit para Adicionar las funciones del caso CU01: Agendar citas de los clientes: Se realiza la creación de la entidad Cita con su Dto, esto con respecto a los atributos que se deben almacenar para tener un agendamiento de citas, ademas se hace la adicion del repositorio de la cita, de las operaciones que se realizan para crear la cita, el  servicio que implementa dichas operaciones y finalmente el controlador de la cita.
De igual manera se hace la creación de las paginas (Html) para el manejo desde la vista de esta funcionalidad creada

+++ src/main/java/co/ucentral/sistemas/proyectoCitasG15/entidadesDto/ClienteDto.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @AllArgsConstructor
 public class ClienteDto {
 
-    private long idCliente;
+    private int idCliente;
     private String nombre;
     private int edad;
     private String cedula;
